Release memory used after calling an unarycall
With this change the memory is apparently released at each unarycall. Has been spotted
that the memory keeps increasing but a very small pace, which can be also reproduced with
the synchronous version but with an even smaller increment.

Now, longer load tests can be executed without consuming the whole memory, even though further
investigations will be needed to address this likely memory leak that might be there from the
very beginning.